+++ src/app/modules/book/frontend/index.ts
@@ -1,0 +1,3 @@
+export * from './Book_List';
+export * from './Book_View';
+export * from './Book_Form';