+++ src/app/modules/catalog/frontend/Catalog_View.tsx
@@ -1,4 +1,4 @@
-import { useContext, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { useParams } from "react-router";
 import { Descriptions } from "antd";
 import {
@@ -6,22 +6,32 @@ import {
   Settings_Request,
 } from "../../../common/types/RequestTypes";
 import { DocCatalogType } from "../../../common/types/DocCatalog";
-import { IPC_DATABASE, IPC_SETTINGS } from "../../../common/types/IPC_Channels";
-import { DOCTYPE_CATALOG } from "../../../common/types/DocType";
+import { IPC_SETTINGS } from "../../../common/types/IPC_Channels";
+import { DocType, DOCTYPE_CATALOG } from "../../../common/types/DocType";
 
 import { App_Context } from "../../../frontend/App_Context";
 import { Header_Buttons_IPC } from "../../../frontend/Header_Buttons_IPC";
 import { App_Messages_IPC } from "../../../frontend/App_Messages_IPC";
+import { modul_props } from "../modul_props";
 
 export function Catalog_View() {
   const { id } = useParams();
-  const app_context = useContext(App_Context);
 
   const [dataObject, setDataObject] = useState<DocCatalogType>(null);
 
+  const doclabel: string = modul_props.doclabel;
+  const doctype: DocType = modul_props.doctype;
+  const segment: string =  modul_props.segment;
+
   useEffect(() => {
-    console.log("ContextData", app_context);
-    Header_Buttons_IPC.request_buttons("view", "catalog", id);
+    Header_Buttons_IPC.request_buttons({
+      viewtype: "view",
+      doctype: doctype,
+      doclabel: doclabel,
+      id: id,
+      surpress: false,
+      options: {},
+    });
 
     const request: Settings_Request = {
       type: "request:get-connection",