+++ src/app/modules/catalog/frontend/index.ts
@@ -1,0 +1,3 @@
+export * from './Catalog_List';
+export * from './Catalog_View';
+export * from './Catalog_Form';