+++ src/app/modules/ipc/modul_props.ts
@@ -1,0 +1,7 @@
+import { Modul_Props_I } from "../../common/Modul_Props";
+
+export const modul_props: Modul_Props_I = {
+  doclabel: "IPC Test",
+  doctype: "ipc",
+  segment: "",
+};