+++ src/app/modules/user/frontend/User_View.tsx
@@ -1,25 +1,33 @@
-import { useContext, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { useNavigate, useParams } from "react-router";
 import { Descriptions } from "antd";
 import { Action_Request, DB_Request } from "../../../common/types/RequestTypes";
 import { DocUserType } from "../../../common/types/DocUser";
 import { IPC_DATABASE } from "../../../common/types/IPC_Channels";
-import { DOCTYPE_USER } from "../../../common/types/DocType";
+import { DocType } from "../../../common/types/DocType";
 
-import { App_Context } from "../../../frontend/App_Context";
 import { Header_Buttons_IPC } from "../../../frontend/Header_Buttons_IPC";
 import { App_Messages_IPC } from "../../../frontend/App_Messages_IPC";
+import { modul_props } from "../modul_props";
 
 export function User_View() {
   const navigate = useNavigate();
   const { id } = useParams();
-  const app_context = useContext(App_Context);
-
   const [dataObject, setDataObject] = useState<DocUserType>(null);
 
+  const doclabel: string = modul_props.doclabel;
+  const doctype: DocType = modul_props.doctype;
+  const segment: string = modul_props.segment;
+  
   useEffect(() => {
-    console.log("ContextData", app_context);
-    Header_Buttons_IPC.request_buttons("view", "user", id);
+    Header_Buttons_IPC.request_buttons({
+      viewtype: "view",
+      doctype: doctype,
+      doclabel: doclabel,
+      id: id,
+      surpress: false,
+      options: {},
+    });
 
     const request: DB_Request = {
       type: "request:data",
@@ -34,7 +42,7 @@ export function User_View() {
         setDataObject(result);
         App_Messages_IPC.request_message(
           "request:message-info",
-          App_Messages_IPC.get_message_from_request(request.type, "User")
+          App_Messages_IPC.get_message_from_request(request.type, doclabel)
         );
       })
       .catch(function (error: any) {
@@ -49,7 +57,7 @@ export function User_View() {
     const buaUnsubscribe = window.electronAPI.listen_to(
       "ipc-button-action",
       (response: Action_Request) => {
-        if (response.target === DOCTYPE_USER && response.view == "view") {
+        if (response.target === doctype && response.view == "view") {
           console.log("User_View says ACTION: ", response);
           App_Messages_IPC.request_message("request:message-info", JSON.stringify(response));
         }