+++ src/app/modules/user/frontend/index.ts
@@ -1,0 +1,3 @@
+export * from './User_List';
+export * from './User_View';
+export * from './User_Form';